Move hashJoin related cascading code into CascadingBackend
